Updating the version of os-core-lib-aws
SNAPSHOTs are not stable for use as a dependency. When this commit was being prepared, the SNAPSHOTs had been recently purged for
recovering disk space. os-core-lib-aws has moved on to version 0.12.0-SNAPSHOT, so rebuilding the old would require diving into the
git history. It is easier to update the pom to use a release version.